server/requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Database
sqlalchemy==2.0.35
alembic==1.13.2
psycopg2-binary==2.9.9

# Validation & Configuration
pydantic-settings==2.5.2

# Google Auth & Drive
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.149.0

# Auth
python-jose[cryptography]==3.3.0

# HTTP
requests==2.32.3
httpx==0.27.2

# Document Parsing
pdfminer.six==20231228
python-pptx==1.0.2
python-docx==1.1.2

# LLM
openai==1.51.0

# Forms
python-multipart==0.0.12

# Environment
python-dotenv==1.0.1
